Update aggregation + pbft support for metadata
